+++ beaker/exceptions.py
@@ -50,6 +50,12 @@ class DatasetNotFound(BeakerError):
     pass
 
 
+class EmptySourceFileError(BeakerError):
+    """
+    Raised when created a dataset from a file when the file is empty.
+    """
+
+
 class JobNotFound(BeakerError):
     pass